refactor(web): consolidate format helpers into lib/format.ts
Absorbs formatParticao and normalizeSearchInput plus the in-component
copies of formatBRL and formatDate that SearchHero and ContractDetailView
kept locally, and the ad-hoc "R$ ${n.toLocaleString('pt-BR')}" ternary
and inline Date toLocaleDateString that Agency/City repeated. Agency
and City contract values now render with currency decimals and
SearchHero dates show em-dash instead of blank — small visual
unification in exchange for a single canonical format module.
